Implement support for application plugins and extensions
