+++ files/pkgconfig/webapp-plugin.pc.in
@@ -4,5 +4,5 @@ includedir=@WEBOS_INSTALL_INCLUDEDIR@
 Name: webapp-plugin
 Description: Library to build native plugins for webOS ports web applications
 Version: @WEBOS_COMPONENT_VERSION@
-Libs: -L${libdir} -l@CMAKE_PROJECT_NAME@
+Libs: -L${libdir} -lwebapp-plugin
 Cflags: -I${includedir} -I${includedir}/webapp-plugin
