+++ src/resources.qrc
@@ -4,5 +4,7 @@
         <file>qml/extensionmanager.js</file>
         <file>qml/webos-api.js</file>
         <file>qml/ApplicationContainer.qml</file>
+        <file>extensions/PalmServiceBridge.js</file>
+        <file>extensions/PalmSystem.js</file>
     </qresource>
 </RCC>
