+++ src/webapplicationplugin.h
@@ -15,14 +15,37 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>
  */
 
-#include "scriptexecutor.h"
+
+#ifndef WEBAPPLICATIONPLUGIN_H
+#define WEBAPPLICATIONPLUGIN_H
+
+#include <QObject>
+#include <QFileInfo>
+#include <QPluginLoader>
+#include <QFileInfo>
+
+#include <applicationplugin.h>
 
 namespace luna
 {
 
-ScriptExecutor::ScriptExecutor(QObject *parent) :
-    QObject(parent)
+class WebApplicationPlugin : public QObject,
+                             public ApplicationPlugin
 {
-}
+    Q_OBJECT
+public:
+    WebApplicationPlugin(const QFileInfo &path, QObject *parent = 0);
+
+    bool load();
+
+    QList<BaseExtension*> createExtensions(ApplicationEnvironment *environment);
+
+private:
+    QPluginLoader mLoader;
+    ApplicationPlugin *mInstance;
+    QFileInfo mPath;
+};
 
 } // namespace luna
+
+#endif // WEBAPPLICATIONPLUGIN_H
